+++ llm_assistant/tests/__init__.py
@@ -7,3 +7,4 @@ from . import test_stream_duration_cap
 from . import test_effective_tools
 from . import test_tool_resolver
 from . import test_prompt_assembler
+from . import test_message_window_pin